Teams_T2_2022/Cloud_Dev/pipeline/requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorlog==6.7.0
cycler==0.11.0
datasets==2.4.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
denoiser==0.1.5
dill==0.3.5.1
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
filelock==3.8.0
Flask
flatbuffers==1.12
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
gast==0.3.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h11==0.13.0
h5py==2.10.0
huggingface-hub==0.9.1
hydra-colorlog==0.1.4
hydra-core==0.11.3
HyperPyYAML==1.0.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.14.0
julius==0.2.6
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
matplotlip==0.2
mistune==2.0.4
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numba
numpy
oauthlib==3.2.0
omegaconf==1.4.1
opencv-python-headless
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow
pkgutil_resolve_name==1.3.10
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pystoi==0.3.3
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.0
responses==0.18.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.1
selenium==4.4.3
semantic-version==2.10.0
Send2Trash==1.8.0
sentencepiece==0.1.97
setuptools-rust==1.5.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
sounddevice==0.4.5
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
speechbrain==0.5.13
stack-data==0.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.11
Wave==0.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.2
wrapt==1.12.1
wsproto==1.1.0
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
